moved types to a components section
