+++ app/src/components/proposals/proposal.types.ts
