+++ app/src/components/proposals/proposalPage/VotingSidebar/votingColumn/StandardVoteCard.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react"
 import VoteButton from "@/components/proposals/proposalPage/VotingSidebar/votingColumn/VoteButton"
-import { VoteType } from "@/app/proposals/proposal.types"
+import { VoteType } from "@/components/proposals/proposal.types"
 
 interface standardVoteCardProps {
   selectedVote: VoteType | null
